+++ pulse-sys/CHANGELOG.md
@@ -1,4 +1,4 @@
-# [unreleased]
+# 1.15.0 (November 25th, 2020)
 
  * Marked `pa_encoding_t` as `#[non_exhaustive]`
  * Added derive of `FromPrimitive` and `ToPrimitive` from the `num-derive` crate on enums